preserve non-pod involved object names
The short_name logic seems to be targeted at pods where the last two
segments are auto generated. In other cases the mangling is undesirable
as it throws away relevant parts of the name when the name contains
dashes. For example a Service called "my-lb" would appear in the culprit
only as "my".